refactor: AdminBoard 구조 변경
