+++ admin/src/main/java/com/kernel/admin/repository/AdminBoardRepository.java
@@ -1,7 +1,7 @@
 package com.kernel.admin.repository;
 
 import com.kernel.admin.domain.entity.Board;
-import com.kernel.admin.domain.enumerate.BoardType;
+import com.kernel.admin.domain.enums.BoardType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
