--- admin/src/main/java/com/kernel/admin/service/banner/UploadedFileRepository.java
@@ -1,7 +1,0 @@
-package com.kernel.admin.service.banner;
-
-import com.kernel.admin.domain.entity.UploadedFiles;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UploadedFileRepository extends JpaRepository<UploadedFiles, Long> {
-}
